// src/core/MainScoreBoard/components/LeagueInfo/models/ILeague.ts
export interface ILeague {
     league: {
          id: 311;
          name: string;
          type: string;
          logo: string;
     };
     country: {
          name: string;
          code: string;
          flag: string;
     };
     seasons: [
          {
               year: number;
               start: string;
               end: string;
               current: boolean;
               coverage: {
                    fixtures: {
                         events: boolean;
                         lineups: boolean;
                         statistics_fixtures: boolean;
                         statistics_players: boolean;
                    };
                    standings: boolean;
                    players: boolean;
                    top_scorers: boolean;
                    top_assists: boolean;
                    top_cards: boolean;
                    injuries: boolean;
                    predictions: boolean;
                    odds: boolean;
               };
          }
     ];
}